+++ web/src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from 'next'
 import { Comfortaa as Font } from 'next/font/google'
 import { NextIntlClientProvider } from 'next-intl'
 import { getMessages } from 'next-intl/server'
+import { cookies } from 'next/headers'
 import { QueryProvider } from '@/providers/query-provider'
 import './globals.css'
 
@@ -23,9 +24,11 @@ export default async function RootLayout({
   children: React.ReactNode
 }>) {
   const messages = await getMessages()
+  const cookieStore = await cookies()
+  const theme = cookieStore.get('theme')?.value
 
   return (
-    <html lang="es">
+    <html lang="es" className={theme === 'dark' ? 'dark' : ''}>
       <body className={`${font.className} antialiased`}>
         <NextIntlClientProvider messages={messages}>
           <QueryProvider>{children}</QueryProvider>
