+++ web/src/components/theme-provider.tsx
@@ -1,0 +1,23 @@
+'use client'
+
+import { useEffect } from 'react'
+
+export function ThemeProvider({ children }: { children: React.ReactNode }) {
+  useEffect(() => {
+    // Apply theme immediately on mount
+    const savedTheme = localStorage.getItem('theme')
+    const systemPrefersDark = window.matchMedia(
+      '(prefers-color-scheme: dark)'
+    ).matches
+    const shouldBeDark =
+      savedTheme === 'dark' || (!savedTheme && systemPrefersDark)
+
+    if (shouldBeDark) {
+      document.documentElement.classList.add('dark')
+    } else {
+      document.documentElement.classList.remove('dark')
+    }
+  }, [])
+
+  return <>{children}</>
+}
